+++ CHANGES.md
@@ -1,6 +1,8 @@
 HEAD
 ----
 * SystemEvents are no longer exceptions. Boo exceptions as flow control!
+* Celluloid::Mailbox#system_event eliminated and replaced with Mailbox#<<
+  SystemEvents are now automatically high priority
 
 0.11.1
 ------